TST/REF: collect index tests (#39485)
