+++ pandas/tests/indexes/numeric/test_setops.py
@@ -110,6 +110,24 @@ class TestSetOps:
             expected = expected.sort_values()
         tm.assert_index_equal(result, expected)
 
+    def test_symmetric_difference(self, sort):
+        # smoke
+        index1 = Index([5, 2, 3, 4], name="index1")
+        index2 = Index([2, 3, 4, 1])
+        result = index1.symmetric_difference(index2, sort=sort)
+        expected = Index([5, 1])
+        assert tm.equalContents(result, expected)
+        assert result.name is None
+        if sort is None:
+            expected = expected.sort_values()
+        tm.assert_index_equal(result, expected)
+
+        # __xor__ syntax
+        with tm.assert_produces_warning(FutureWarning):
+            expected = index1 ^ index2
+        assert tm.equalContents(result, expected)
+        assert result.name is None
+
 
 class TestSetOpsSort:
     @pytest.mark.parametrize("slice_", [slice(None), slice(0)])
