+++ pandas/tests/indexes/test_any_index.py
@@ -3,6 +3,8 @@ Tests that can be parametrized over _any_ Index object.
 
 TODO: consider using hypothesis for these.
 """
+import re
+
 import pytest
 
 import pandas._testing as tm
@@ -33,10 +35,25 @@ def test_mutability(index):
         index[0] = index[0]
 
 
+def test_map_identity_mapping(index):
+    # GH#12766
+    tm.assert_index_equal(index, index.map(lambda x: x))
+
+
 def test_wrong_number_names(index):
     names = index.nlevels * ["apple", "banana", "carrot"]
     with pytest.raises(ValueError, match="^Length"):
         index.names = names
+
+
+def test_view_preserves_name(index):
+    assert index.view().name == index.name
+
+
+def test_ravel_deprecation(index):
+    # GH#19956 ravel returning ndarray is deprecated
+    with tm.assert_produces_warning(FutureWarning):
+        index.ravel()
 
 
 class TestConversion:
@@ -77,10 +94,27 @@ class TestRoundTrips:
             # GH#8367 round-trip with timezone
             assert index.equal_levels(result)
 
+    def test_pickle_preserves_name(self, index):
+        original_name, index.name = index.name, "foo"
+        unpickled = tm.round_trip_pickle(index)
+        assert index.equals(unpickled)
+        index.name = original_name
+
 
 class TestIndexing:
     def test_slice_keeps_name(self, index):
         assert index.name == index[1:].name
+
+    @pytest.mark.parametrize("item", [101, "no_int"])
+    # FutureWarning from non-tuple sequence of nd indexing
+    @pytest.mark.filterwarnings("ignore::FutureWarning")
+    def test_getitem_error(self, index, item):
+        msg = r"index 101 is out of bounds for axis 0 with size [\d]+|" + re.escape(
+            "only integers, slices (`:`), ellipsis (`...`), numpy.newaxis (`None`) "
+            "and integer or boolean arrays are valid indices"
+        )
+        with pytest.raises(IndexError, match=msg):
+            index[item]
 
 
 class TestRendering:
