+++ pandas/tests/indexes/test_indexing.py
@@ -16,12 +16,16 @@ contain tests for the corresponding methods specific to those Index subclasses.
 import numpy as np
 import pytest
 
+from pandas.errors import InvalidIndexError
+
 from pandas import (
     DatetimeIndex,
     Float64Index,
     Index,
     Int64Index,
+    IntervalIndex,
     PeriodIndex,
+    Series,
     TimedeltaIndex,
     UInt64Index,
 )
@@ -135,6 +139,62 @@ class TestContains:
         assert 1.1 in float_index
         assert 1.0 not in float_index
         assert 1 not in float_index
+
+
+class TestGetValue:
+    @pytest.mark.parametrize(
+        "index", ["string", "int", "datetime", "timedelta"], indirect=True
+    )
+    def test_get_value(self, index):
+        # TODO: Remove function? GH#19728
+        values = np.random.randn(100)
+        value = index[67]
+
+        with pytest.raises(AttributeError, match="has no attribute '_values'"):
+            # Index.get_value requires a Series, not an ndarray
+            with tm.assert_produces_warning(FutureWarning):
+                index.get_value(values, value)
+
+        with tm.assert_produces_warning(FutureWarning):
+            result = index.get_value(Series(values, index=values), value)
+        tm.assert_almost_equal(result, values[67])
+
+
+class TestGetIndexer:
+    def test_get_indexer_consistency(self, index):
+        # See GH#16819
+        if isinstance(index, IntervalIndex):
+            # requires index.is_non_overlapping
+            return
+
+        if index.is_unique:
+            indexer = index.get_indexer(index[0:2])
+            assert isinstance(indexer, np.ndarray)
+            assert indexer.dtype == np.intp
+        else:
+            e = "Reindexing only valid with uniquely valued Index objects"
+            with pytest.raises(InvalidIndexError, match=e):
+                index.get_indexer(index[0:2])
+
+        indexer, _ = index.get_indexer_non_unique(index[0:2])
+        assert isinstance(indexer, np.ndarray)
+        assert indexer.dtype == np.intp
+
+
+class TestConvertSliceIndexer:
+    def test_convert_almost_null_slice(self, index):
+        # slice with None at both ends, but not step
+
+        key = slice(None, None, "foo")
+
+        if isinstance(index, IntervalIndex):
+            msg = "label-based slicing with step!=1 is not supported for IntervalIndex"
+            with pytest.raises(ValueError, match=msg):
+                index._convert_slice_indexer(key, "loc")
+        else:
+            msg = "'>=' not supported between instances of 'str' and 'int'"
+            with pytest.raises(TypeError, match=msg):
+                index._convert_slice_indexer(key, "loc")
 
 
 @pytest.mark.parametrize(
